refator: refactored components by grouping them by pages
