+++ frontend/src/App.css
@@ -28,6 +28,18 @@
   color: #61dafb;
 }
 
+.units {
+  display: flex;
+  flex-wrap: wrap;
+}
+
+.unit {
+  width: calc(100% / 3);
+  height: auto;
+  object-fit: cover;
+  object-position: center;
+}
+
 /* @keyframes App-logo-spin {
   from {
     transform: rotate(0deg);
