+++ frontend/src/components/groupsPage/GroupCard.jsx
