+++ frontend/src/components/studentsPage/StudentRowStudentDisplay.jsx
@@ -1,8 +1,8 @@
 import { Tr, Td, useDisclosure, useToast } from '@chakra-ui/react';
 import { DeleteIcon } from '@chakra-ui/icons';
 import { useParams } from 'react-router';
-import getToastSettings from './ToastSettings';
-import { MockAuth } from '../mockAuth/mockAuth';
+import getToastSettings from '../ToastSettings';
+import { MockAuth } from '../../mockAuth/mockAuth';
 import { useAuth0 } from '@auth0/auth0-react';
 
 const StudentRowStudentDisplay = ({ studentData, numberOfGroups, onDelete }) => {
