+++ frontend/src/components/unitHomePage/UnitCard.jsx
@@ -23,10 +23,10 @@ import {
     useToast,
 } from '@chakra-ui/react';
 import { BsTrash } from 'react-icons/bs';
-import getToastSettings from './ToastSettings'
+import getToastSettings from '../ToastSettings'
 import { IoEllipsisHorizontalSharp } from 'react-icons/io5';
 import { BrowserRouter, Routes, Route, useNavigate } from 'react-router-dom';
-import { MockAuth } from '../mockAuth/mockAuth';
+import { MockAuth } from '../../mockAuth/mockAuth';
 import { useAuth0 } from '@auth0/auth0-react';
 
 const UnitCard = (unit) => {
