--- frontend/src/pages/UnitHomePage.css
@@ -1,11 +1,0 @@
-.units {
-  display: flex;
-  flex-wrap: wrap;
-}
-
-.unit {
-  width: calc(100% / 3);
-  height: auto;
-  object-fit: cover;
-  object-position: center;
-}
