+++ frontend/src/pages/UnitHomePage.jsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useAuth0 } from '@auth0/auth0-react';
-import UnitCard from '../components/UnitCard';
-import '../pages/UnitHomePage.css';
+import UnitCard from '../components/unitHomePage/UnitCard';
+import '../App.css';
 import getToastSettings from '../components/ToastSettings';
 import { MockAuth } from '../mockAuth/mockAuth';
